backend/requirements.txt
ghunt
naminter
fastapi
uvicorn
sherlock-project
holehe
google-genai
python-dotenv
